能用枚举的尽量用枚举，提高可读性

+++ filecoin4j-rpc/src/main/java/com/github/mufanh/filecoin4j/domain/types/BlockHeader.java
@@ -28,7 +28,7 @@ public class BlockHeader implements Serializable {
 
     private List<Cid> parents; // 5
 
-    private Long parentWeight; // 6
+    private BigInteger parentWeight; // 6
 
     private Long height; // 7
